Fix infinite recursion when grammar contains a class "Link"; fixes #3
The problem occurs because ``displayResult`` uses the class name
``schema_link`` for anchors linking to result schema, which would then
also contain ``schema_link`` class indicating an element "is a schema
of link".

This renames the class used for elements containing "schema of STH" to
``schema_of_NAME`` which avoids the conflict.