Workaround lack of namespace axis support in Firefox
See Firefox bug [#94270](https://bugzilla.mozilla.org/show_bug.cgi?id=94270)
for details, I based the workaround on the code provided in
[comment #28](https://bugzilla.mozilla.org/show_bug.cgi?id=94270#c28).

Basically because Firefox lacks support for ``namespace:`` axis,
``$ns-uri` in ``getHyperlinkedElement`` ends up being null even for
elements that have namespace (e.g. ``xs:int``).  This causes the
``xsl:choose`` element to fall back to ``when test="$qname"`` and the
elements are displayed as e.g. 'xs:int' as schema_link that does
nothing.

This adds a check to see if there are no namespace elements but the
element has a prefix, in which case it will just display it as plain
local name, e.g. ``int``.  Still using an <a> tag for it to keep it in
the same style as non-namespaced elements.