Fix weird issues with schema links in chrome
Fix issue where a link to schema for complex elements may not be
generated correctly (in chrome).

Turns out ``boolean($ns-uri)`` is actually true when ``$ns-uri = ''``
in many XSLT engines including the one used in chrome (it is false in
FF) (as used in test:

    $ns-uri and starts-with($ns-uri, 'http://www.w3.org/XML/') = false

inside the ``getHyperlinkedElement`` template).

This fixes the problem by changing this test to:

    not(normalize-space($ns-uri) = '') and starts-with($ns-uri, 'http://www.w3.org/XML/') = false

which is safer and seems to work consistently across browsers (at least
FF and Chrome).